Separate by models/scrapers not by companies/internships
